Move forge config tests to integration
commit-id:76c69559
